requirements.txt
google-genai
streamlit
pdfplumber
pandas